requirements.txt
Flask==1.1.2
PyYAML==5.4.1
gunicorn==20.1.0
